+++ frontend/src/App.css
@@ -10,7 +10,8 @@
   top: 0;
   left: 0;
   width: 100%;
-  height: 100%;
+  height: 100vh;
+  z-index: -1;
 }
 
 .background-layer {
@@ -19,7 +20,11 @@
   left: 0;
   width: 100%;
   height: 100%;
+  background-size: cover;
+  background-position: center;
+  background-repeat: no-repeat;
   transition: opacity 0.6s ease-in-out;
+  will-change: opacity;
 }
 
 .background-layer.fade-out {
