+++ frontend/src/components/character/InventoryInfo.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import type { InventoryItem } from "../types";
+import type { InventoryItem } from "../../types";
 
 interface InventoryInfoProps {
   inventory: InventoryItem[];
