+++ frontend/src/components/character/index.ts
@@ -1,0 +1,2 @@
+export { default as CharacterInfo } from "./CharacterInfo";
+export { default as InventoryInfo } from "./InventoryInfo";
