+++ frontend/src/components/game/GameEndedScreen.tsx
