+++ frontend/src/components/game/index.ts
@@ -1,0 +1,4 @@
+export { default as GameLayout } from "./GameLayout";
+export { default as StoryContent } from "./StoryContent";
+export { default as GameEndedScreen } from "./GameEndedScreen";
+export { default as Background } from "./Background";
