+++ frontend/src/components/index.ts
@@ -1,0 +1,16 @@
+// Game components
+export { GameLayout, StoryContent, GameEndedScreen, Background } from "./game";
+
+// UI components
+export { Header, Footer, HomeIntroSection } from "./ui";
+
+// Character components
+export { CharacterInfo, InventoryInfo } from "./character";
+
+// Save/Load components
+export { SaveLoadControls, LoadGameModal } from "./save-load";
+
+// Common components
+export { LoadingSpinner } from "./common";
+
+//Made a separate index,ts file to keep import statements clean
