+++ frontend/src/components/save-load/index.ts
@@ -1,0 +1,2 @@
+export { default as SaveLoadControls } from "./SaveLoadControls";
+export { default as LoadGameModal } from "./LoadGameModal";
