+++ frontend/src/components/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as Header } from "./Header";
+export { default as Footer } from "./Footer";
+export { default as HomeIntroSection } from "./HomeIntroSection";
