+++ frontend/src/hooks/game/index.ts
@@ -1,0 +1,5 @@
+export { useGameState } from "./useGameState";
+export { useChoiceManagement } from "./useChoiceManagement";
+export { useInventoryManagement } from "./useInventoryManagement";
+export { useGameRestart } from "./useGameRestart";
+export { useGamePersistence } from "./useGamePersistence";
