+++ frontend/src/pages/LoginPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
+import Header from "../components/ui/Header";
+import Footer from "../components/ui/Footer";
 import { login } from "../api/auth";
 
 const LoginPage: React.FC = () => {
